deduplicate code & fix typo
