+++ app/src/main/java/com/alphawallet/app/entity/tokens/ERC721Ticket.java
@@ -3,10 +3,7 @@ package com.alphawallet.app.entity.tokens;
 import android.content.Context;
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.util.Base64;
 import android.view.View;
-import android.widget.ProgressBar;
-import android.widget.TextView;
 
 import com.alphawallet.app.R;
 import com.alphawallet.app.entity.ContractType;
@@ -20,25 +17,17 @@ import com.alphawallet.app.repository.entity.RealmToken;
 import com.alphawallet.app.service.AssetDefinitionService;
 import com.alphawallet.app.ui.widget.holder.TokenHolder;
 import com.alphawallet.app.viewmodel.BaseViewModel;
-import com.alphawallet.app.web3.Web3TokenView;
 import com.alphawallet.token.entity.TicketRange;
-import com.alphawallet.token.entity.TokenScriptResult;
-import com.alphawallet.token.tools.TokenDefinition;
 
 import org.web3j.abi.datatypes.Function;
 import org.web3j.abi.datatypes.generated.Uint256;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.Locale;
-
-import io.reactivex.android.schedulers.AndroidSchedulers;
-import io.reactivex.schedulers.Schedulers;
 
 public class ERC721Ticket extends Token implements Parcelable {
 
@@ -181,45 +170,6 @@ public class ERC721Ticket extends Token implements Parcelable {
         tokenHolder.balanceEth.setText(composite);
 
         tokenHolder.layoutValueDetails.setVisibility(View.GONE);
-    }
-
-    /*************************************
-     *
-     * Conversion functions used for manipulating indices
-     *
-     */
-
-    /**
-     * This is a single method that populates any instance of graphic ticket anywhere
-     *
-     * @param range
-     * @param activity
-     * @param assetService
-     * @param ctx needed to create date/time format objects
-     */
-    @Override
-    public void displayTicketHolder(TicketRange range, View activity, AssetDefinitionService assetService, Context ctx, boolean iconified)
-    {
-        TokenDefinition td = assetService.getAssetDefinition(tokenInfo.chainId, tokenInfo.address);
-        if (td != null)
-        {
-            //use webview
-            displayTokenscriptView(range, assetService, activity, ctx, iconified);
-        }
-        else
-        {
-            activity.findViewById(R.id.layout_legacy).setVisibility(View.VISIBLE);
-            activity.findViewById(R.id.layout_webwrapper).setVisibility(View.GONE);
-
-            TextView amount = activity.findViewById(R.id.amount);
-            TextView name = activity.findViewById(R.id.name);
-
-            String nameStr = getTokenTitle();
-            String seatCount = String.format(Locale.getDefault(), "x%d", range.tokenIds.size());
-
-            name.setText(nameStr);
-            amount.setText(seatCount);
-        }
     }
 
     public void checkIsMatchedInXML(AssetDefinitionService assetService)
