+++ app/src/main/java/com/alphawallet/app/entity/tokens/Token.java
@@ -8,6 +8,7 @@ import android.text.TextUtils;
 import android.util.Base64;
 import android.view.View;
 import android.widget.ProgressBar;
+import android.widget.TextView;
 
 import com.alphawallet.app.R;
 import com.alphawallet.app.entity.ContractType;
@@ -24,6 +25,7 @@ import com.alphawallet.app.viewmodel.BaseViewModel;
 import com.alphawallet.app.web3.Web3TokenView;
 import com.alphawallet.token.entity.TicketRange;
 import com.alphawallet.token.entity.TokenScriptResult;
+import com.alphawallet.token.tools.TokenDefinition;
 
 import org.web3j.abi.datatypes.Function;
 import org.web3j.utils.Numeric;
@@ -36,6 +38,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -604,11 +607,6 @@ public class Token implements Parcelable, Comparable<Token>
     public void checkIsMatchedInXML(AssetDefinitionService assetService) { }
     public int[] getTicketIndices(String ticketIds) { return new int[0]; }
     public boolean contractTypeValid() { return !(contractType == ContractType.NOT_SET || contractType == ContractType.OTHER); }
-
-    public void displayTicketHolder(TicketRange range, View activity, AssetDefinitionService assetService, Context ctx, boolean iconified) { }
-    public void displayTicketHolder(TicketRange range, View activity, AssetDefinitionService assetService, Context ctx) {
-        displayTicketHolder(range, activity, assetService, ctx, false);
-    }
     public List<BigInteger> getArrayBalance() { return new ArrayList<>(); }
     public List<BigInteger> getNonZeroArrayBalance() { return new ArrayList<>(Arrays.asList(BigInteger.ZERO)); }
     public boolean isMatchedInXML() { return false; }
@@ -1021,6 +1019,55 @@ public class Token implements Parcelable, Comparable<Token>
     /**
      * Common TokenScript methods
      */
+
+    public void displayTicketHolder(TicketRange range, View activity, AssetDefinitionService assetService, Context ctx) {
+        displayTicketHolder(range, activity, assetService, ctx, false);
+    }
+
+    /**
+     * This is a single method that populates any instance of graphic ticket anywhere
+     *
+     * @param range
+     * @param activity
+     * @param assetService
+     * @param ctx needed to create date/time format objects
+     */
+    public void displayTicketHolder(TicketRange range, View activity, AssetDefinitionService assetService, Context ctx, boolean iconified)
+    {
+        //need to wait until the assetDefinitionService has finished loading assets
+        assetService.getAssetDefinitionASync(tokenInfo.chainId, tokenInfo.address)
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(td -> renderTicketHolder(td, range, activity, assetService, ctx, iconified), this::loadingError).isDisposed();
+    }
+
+    private void loadingError(Throwable e)
+    {
+        e.printStackTrace();
+    }
+
+    private void renderTicketHolder(TokenDefinition td, TicketRange range, View activity, AssetDefinitionService assetService, Context ctx, boolean iconified)
+    {
+        if (td != null && td.holdingToken != null)
+        {
+            //use webview
+            displayTokenscriptView(range, assetService, activity, ctx, iconified);
+        }
+        else
+        {
+            activity.findViewById(R.id.layout_legacy).setVisibility(View.VISIBLE);
+            activity.findViewById(R.id.layout_webwrapper).setVisibility(View.GONE);
+
+            TextView amount = activity.findViewById(R.id.amount);
+            TextView name = activity.findViewById(R.id.name);
+
+            String nameStr = getTokenTitle();
+            String seatCount = String.format(Locale.getDefault(), "x%d", range.tokenIds.size());
+
+            name.setText(nameStr);
+            amount.setText(seatCount);
+        }
+    }
 
     protected void displayTokenscriptView(TicketRange range, AssetDefinitionService assetService, View activity, Context ctx, boolean iconified)
     {
